# buzzbar-backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: BuzzBar Backend API
  version: "1.0.0"
  description: |
    Dev Phase 1 API (stable).

    Stability guarantee:
    - `/api/v1/**` is considered stable after Phase 1.
    - Future changes must be additive only. Breaking changes require `/api/v2`.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/v1/admin/payments:
    get:
      tags: [AdminPayments]
      summary: List payment transactions
      description: |
        Admin/SuperAdmin-only payments list used by the operational payments module.
        Supports server-side filtering, sorting, pagination, and order/user search.
        `q` matches payment transaction id, provider reference, order number, and user identity fields.
        `stalePending=true` limits results to wallet transactions still pending beyond the configured timeout window.
      security: [{ AdminBearerAuth: [] }]
      parameters:
        - in: query
          name: provider
          required: false
          schema: { type: string, enum: [MOCK, ESEWA, KHALTI] }
        - in: query
          name: status
          required: false
          schema: { type: string, enum: [INITIATED, PENDING, SUCCESS, FAILED] }
        - in: query
          name: paymentMethod
          required: false
          schema: { type: string, enum: [COD, WALLET] }
        - in: query
          name: from
          required: false
          schema: { type: string, format: date-time }
        - in: query
          name: to
          required: false
          schema: { type: string, format: date-time }
        - in: query
          name: q
          required: false
          schema: { type: string }
        - in: query
          name: stalePending
          required: false
          schema: { type: boolean }
        - in: query
          name: sort
          required: false
          schema: { type: string, enum: [createdAt_desc, createdAt_asc, amount_desc, amount_asc, updatedAt_desc], default: createdAt_desc }
        - in: query
          name: page
          required: false
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          required: false
          schema: { type: integer, enum: [20, 50, 100], default: 20 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          type: object
                          properties:
                            _id: { type: string }
                            provider: { type: string, enum: [MOCK, ESEWA, KHALTI] }
                            paymentMethod: { type: string, enum: [COD, WALLET] }
                            status: { type: string, enum: [INITIATED, PENDING, SUCCESS, FAILED] }
                            amount: { type: integer }
                            currency: { type: string }
                            providerReference: { type: string, nullable: true }
                            createdAt: { type: string, format: date-time }
                            updatedAt: { type: string, format: date-time }
                            isMock: { type: boolean }
                            finality: { type: string, enum: [FINAL, OPEN] }
                            stalePending: { type: boolean }
                            order:
                              type: object
                              nullable: true
                              properties:
                                id: { type: string, nullable: true }
                                orderNumber: { type: string, nullable: true }
                                status: { type: string, nullable: true }
                                paymentStatus: { type: string, nullable: true }
                                total: { type: integer, nullable: true }
                            user:
                              type: object
                              nullable: true
                              properties:
                                id: { type: string, nullable: true }
                                email: { type: string, nullable: true }
                                phone: { type: string, nullable: true }
                                name: { type: string, nullable: true }
                      page: { type: integer }
                      limit: { type: integer }
                      total: { type: integer }
        "400":
          description: INVALID_LIMIT | INVALID_DATE
        "403":
          description: ADMIN_FORBIDDEN
  /api/v1/admin/payments/{id}:
    get:
      tags: [AdminPayments]
      summary: Get payment transaction
      description: |
        Admin/SuperAdmin-only structured payment detail read model.
        Includes linked order/user context, normalized request/response snapshots,
        failure diagnostics, and mock lifecycle data when the provider is `MOCK`.
      security: [{ AdminBearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      payment:
                        type: object
                        properties:
                          id: { type: string }
                          provider: { type: string, enum: [MOCK, ESEWA, KHALTI] }
                          paymentMethod: { type: string, enum: [COD, WALLET] }
                          status: { type: string, enum: [INITIATED, PENDING, SUCCESS, FAILED] }
                          amount: { type: integer }
                          currency: { type: string }
                          providerReference: { type: string, nullable: true }
                          failureReason: { type: string, nullable: true }
                          createdAt: { type: string, format: date-time }
                          updatedAt: { type: string, format: date-time }
                          isMock: { type: boolean }
                          isFinal: { type: boolean }
                      order:
                        type: object
                        nullable: true
                        properties:
                          id: { type: string }
                          orderNumber: { type: string }
                          status: { type: string }
                          paymentMethod: { type: string, enum: [COD, WALLET] }
                          paymentStatus: { type: string }
                          total: { type: integer }
                          createdAt: { type: string, format: date-time, nullable: true }
                      user:
                        type: object
                        nullable: true
                        properties:
                          id: { type: string }
                          email: { type: string, nullable: true }
                          phone: { type: string, nullable: true }
                          name: { type: string, nullable: true }
                      snapshots:
                        type: object
                        properties:
                          request: { nullable: true }
                          response: { nullable: true }
                      diagnostics:
                        type: object
                        properties:
                          failureReason: { type: string, nullable: true }
                          requestId: { type: string, nullable: true }
                          operatorHint: { type: string, nullable: true }
                          stalePending: { type: boolean }
                          pendingAgeMinutes: { type: integer, nullable: true }
                          providerResult: { type: string, nullable: true }
                          mockLifecycle:
                            type: object
                            nullable: true
                            properties:
                              providerPath: { type: string }
                              steps:
                                type: array
                                items:
                                  type: object
                                  properties:
                                    id: { type: string, enum: [INIT_REQUESTED, INIT_SUCCESS, CONFIRM_REQUESTED, CONFIRM_PENDING, CONFIRM_SUCCESS, CONFIRM_FAILED] }
                                    label: { type: string }
                                    at: { type: string, format: date-time }
                                    state: { type: string, enum: [done, pending, failed] }
        "404":
          description: NOT_FOUND
        "403":
          description: ADMIN_FORBIDDEN
components:
  securitySchemes:
    UserBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...